requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.30.0


# Conservative Django
django==3.2.18

# Configuration
django-environ==0.9.0
django-rest-captcha==0.2.0
django-debug-toolbar==3.8.1
django-extensions==3.2.1
django-model-utils==4.3.1

drf-yasg==1.21.5
drf-openapi==1.3.0

# OPS management
fabric==2.5.0
pyasn==1.6.1

pandas==1.5.3
numpy==1.24.2
openpyxl==3.1.2
